refactor: move util from @blocksuite/editor to @blocksuite/blocks
