+++ packages/blocks/src/index.ts
@@ -17,6 +17,7 @@ export * from './__internal__/rich-text/rich-text-operations.js';
 export { getServiceOrRegister } from './__internal__/service.js';
 export type { BaseService } from './__internal__/service/index.js';
 export * from './__internal__/utils/common-operations.js';
+export * from './__internal__/utils/filesys.js';
 export * from './__internal__/utils/lit.js';
 export * from './__internal__/utils/query.js';
 export * as SelectionUtils from './__internal__/utils/selection.js';
